requirements.txt
asgiref==3.2.5
Django==3.0.4
django-model-utils==4.0.0
djangorestframework==3.12.1
pytz==2019.3
sqlparse==0.3.1
